delete policy applied to controller

+++ app/Http/Controllers/CommentController.php
@@ -83,6 +83,7 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        $this->authorize("delete",$comment);
         $comment->delete();
         return back();
     }
